pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embedding-api"
version = "0.1.0"
description = "Qwen3 Embedding API Service"
authors = [
    { name = "waveman800", email = "xiaoning800@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.0",
    "sentence-transformers>=2.7.0",
    "tiktoken>=0.5.1",
    "uvicorn[standard]>=0.23.2",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.3.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.290",
]

[project.scripts]
embedding-api = "embedding_api.main:app"

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
]

[project.urls]
"Homepage" = "https://github.com/waveman800/embedding_api"
"Bug Tracker" = "https://github.com/waveman800/embedding_api/issues"
